Remove LocalTransactionService
This functionality has been moved into the LocalTransactionPresenter. This was suggested by Kevin Dew in code review as most of LocalTransactionService concerns seem to be based around unavailable handling, with much of it becoming redundant by this work.

As the devolved administration availability related methods determine how information is presented to a user on the view, a presenter seemed like a better place for this functionality.

As a result LocalTransactionService is being removed.