Remove the remaining bespoke code for rendering results
The 'bespoke' code was added to this application to display
different results pages for devolved administrations for some
of the COVID services.

Work is in progress to make specifying devolved administration
availability controllable by content designers using the
Publisher application. [This Trello card](https://trello.com/c/Kgr3I4Tz/401-store-devolved-administration-local-transaction-data-in-model-and-govuk-content-schema) gives further details. As a result,
this bespoke code should now be removed.