Dedupe JavaScript to stop it being loaded twice
This removes the JavaScript that is already loaded by Static. This
duplication meant that the search toggle wasn't working, as the script
was being initialised twice. This meant that a single click would open
and the instantly close the search.

Tests dependencies have been updated to ensure that the JavaScript that
is loaded by Static is still available for the tests in Frontend.