Remove redirect_path and level_of_authentication from sign in
Originally we planned for this to be the only place which construct an
OAuth auth URI, and other apps which need to sign the user in would
redirect via https://www.gov.uk/sign-in?redirect_path=...

We've since changed that approach, with other apps calling account-api
to get a sign in URI directly.  So we can remove these parameters, as
they're unused.