Fix :focus:hover link colour for homepage links
The "Popular on GOV.UK" links on the homepage have the govuk grey colour
on hover, which is adequate in terms of minimum contrast requirements
against the black background.
This text colour is not adequate, however, against the yellow background
of the :focus state. This makes the CSS more specific, so that the grey
colour change only applies when the element is hovered, but not in
focus.