Popular links CSS states adjustments
The "Popular links" on the homepage are over a blue background on
mobile, and a black background otherwise. The text has a grey colour on
the active state, which does not meet minimum contrast requirements
against the blue background. The minimum contrast requirement is also
not achieved in the :active:focus state as the text is grey over a
yellow background (this is only an issue in some browsers, Chrome being
one). This aims to fix these issues.