Do not use Addressable::URI to manipulate querystring
Addressable::URI doesn't handle repeated parameters, which we often
get in the Transition Checker:

    Addressable::URI.parse("https://www.foo.com?a=1&a=2").query_values
    => {"a"=>"2"}

So this means we've been redirecting users back to the wrong results
page from the post-registration confirmation page, only showing
the *last* criterion.  We're still saving the right results, as that's
done as part of the registration journey, but that is going away
imminently, and so this is about to become a more significant problem.