Adds country name for Northern Ireland
The missing country name was causing a number of errors in Sentry [1],
leading to the Northern Ireland page [2] returning a generic GOV.UK 503
page:

```
We're experiencing technical difficulties. Please try again later.
```

The change that caused this bug seems to be when we added functionality
for determining a devolved administration availability [3].

[1]: https://sentry.io/organizations/govuk/issues/2535481416/events/latest/?environment=production&project=202225&referrer=slack
[2]: https://www.gov.uk/contact-electoral-registration-office/electoral-office-for-northern-ireland
[3]: https://github.com/alphagov/frontend/commit/3d076052f5a49480fe006458d4cf76b0f2cd5031#diff-8157218d73a5aeabab17ed84e0cf19d850abd378b8c1acccb0bbe80f302e4f31R33-R35