Use new account layout wrapper for account home
A new and improved layout template was added, with the intention of
being used on pages that should have the account layout, while not being
rendered by the govuk-account-manager-prototype app.

The improvement in this new template compared to the old one is that all
the page furniture for accounts (including the account nav, account
feedback footer, account grid layout) is now contained in the template.
We can remove these elements, and the content associated with them, from
the frontend app, and simply request the gem_layout_account_manager
template, without any visual differences taking place.