Return AB response headers with all variants
At present we're only returning Vary response headers when we get a B variant.
We want to do this all the time or it's confusing for Fastly to know how to cache.

I've added tests for the other variants to make sure it's working everywhere.

The template assertion wasn't working - it was doing an `assert` on the
template name and ignoring the second param. I've changed that to an `assert_equals`
and made the test work.