Remove jQuery from save-location-spec.js
I've done this because I felt that this test was conflated with rather a
lot of wrapping and unwrapping of jQuery objects that I felt was a
little confusing. It isn't essential to remove jQuery from jasmine tests
as Jasmine provides it's own jQuery, however I feel that the use of
jQuery in them can often make understanding JS patterns a bit harder.