Port transactions.js to distinct modules
This takes the transactions.js file, which performed two distinct tasks
(tracking tab interactions, JS enhancements to the form) and splits them
into individual modules to perform both of these actions.

As part of this jQuery is removed from the files and tests are either
port or added (not this code was tested).