Sign in to a service page
Build custom frontend for the sign in to a service (aka sorting hat)
page. The purpose of this page is to help users navigate to some of the
most popular government services.

The HTML for most of the page is defined in the content item so it can
appear in search. The content has been added to account-api together
with a rake task to publish the page.

A notable feature of the page is the search form at the  bottom.
This search form is a regular site search, with the notable difference
that results will be filtered by content types "services" and
"guidance_and_regulation".