Persist cookie_consent flag through OAuth redirects
This is mostly a straight-forward replacement of `redirect_with_ga`
with `redirect_with_analytics`, but in the callback controller we also
need to account for a consent coming from the account-api (which is
how the account-manager currently passes its consent through).

The behaviour I've gone for is to use the consent from the account-api
if present and fall back to the query param otherwise, because the
account-api doesn't give a value when using Digital Identity.