Add Dockerfile specifically for Kubernetes
This commit adds a new Dockerfile specifically targeted at Kubernetes, which switches from the use of Foreman for running Rails to instead running Rails directly. Foreman isn't necessary when running on Kubernetes, as the controller managers within Kubernetes will take care of ensuring that our app is running without the need of a separate process manager.

The reason behind creating a separate Dockerfile in this commit is because the existing Dockerfile is used by the publishing end-to-end tests; once these tests are deprecated (as is the long-term plan) then `Dockerfile.k8s` can become `Dockerfile`.

Co-authored-by: Chris Banks <chris.banks@digital.cabinet-office.gov.uk>