Show cookie & feedback consent page if unset
This commit makes the existing "your information" page appear when a
user authenticates if all of the following are true:

1. account-api returns a `cookie_consent` from the callback
2. account-api returns a `feedback_consent` from the callback
3. At least one of those fields is `nil`

So, this change will currently be a no-op in production, because:

- The PR which makes account-api return `feedback_consent` hasn't
been deployed yet (which makes condition 2 fail)

- We ask those questions in registration in the account-manager, and
send them over to the account-api before redirecting the user back, so
they won't be `nil` (which makes condition 3 fail)

So this change will only rake effect when we switch to Digital
Identity's auth service and users start to register through that, as
they will *not* be asking the consent questions.

A user doesn't get properly logged in (we don't set the session
cookie) until they have provided valid ("yes" or "no") consent to both
things.  We auto-select the appropriate cookie consent radio button by
using the consent returned from account-api (if we end up in some
strange situation where a user has a cookie consent saved but not a
feedback consent), or the `cookie_consent` query parameter from DI
otherwise.