Treat licensing errors as errors
In 7451e52 (back in 2012) we decided to swallow licensing errors so
"frontend still works".

This means that if a user goes to (say) https://www.gov.uk/premises-licence/stoke-on-trent
and there's a transient error with Licensing for whatever reason (a
timeout, or any other kind of non-success response) they don't see an
error page. Instead, they see a page saying:

> You can't apply for this licence online. Contact your local council.

But this message isn't actually true in lots of cases. If the licensing
API times out, that doesn't mean "you can't apply for this licence
online".

To make matters much worse, we serve a Cache-Contol header which says
this response is publicly cacheable for 30 minutes. So in effect we're
magnifying licensing errors - any time there's an error, the Fastly
cache in the nearest POP gets poisoned with an incorrect response for
half an hour. So we've gone from a single error to a local outage.

I think we should just let these errors pass through. Users will
see more "Something went wrong" pages, but that's much better than them
seeing an incorrect and misleading response.

We could instead consider just serving no-cache headers when there's an
error which we're swallowing. But that still leaves users with
misleading responses.