Update stored cookie consent if it doesn't match query param
After talking with DI, we think this is probably the best approach for
now to make the main GOV.UK cookies page integrate nicely with the
account.

Users can use the account-specific cookie consent page to update their
preferences after logging in, but this approach lets users do so
before logging in as well.

If the user has not engaged with the cookie banner, we use the consent
stored in their account, as before.