Remove jQ from filterListItems
Setting the variables `i` and `listItem` to be null stops the linter from complaining. This is necessary because the linter doesn't like variable names to be reused but there isn't a good alternative to `listItem` when working through a collection called `listItems` and we need to iterate through the same list multiple times. We realise that this isn't ideal and are open to alternatives!

Also introduces an `itemsShowing` variable as an easier way to track the number of visible countries

Co-authored-by: Graham Lewis <graham.lewis@digital.cabinet-office.gov.uk>