Apply account layout template to consent page
We need the "Control how we use information about you" page to use the
account template to make for a seamless experience when users are
navigating to and from DI pages and our pages.

This applies the gem_layout_account_manager_no_nav template to this
page, which applies the Account layout header, footer and feedback
message at the top.