Update account home page for email subscriptions
So it shows users' information about their email subscriptions, if they
have any, as well as their Brexit checker results.

Previously, the only service users could use their GOV.UK account for
was the Brexit checker. If they hadn't saved their checker results to
their account, the home page showed a prompt to use the Brexit checker.

Now users will also be able to sign up for email notifications with
their GOV.UK account so it's not correct to show the "You have not used
your account to access any services yet" message if they've not used the
Brexit checker.

We are treating email subscriptions in a slightly different way to other
services that might use the GOV.UK account, which is why I've added a new
helper method to check if a user has subscriptions rather than using the
existing `has_used?`