Fix leaky locale assignment
The Rails locale is not a request-local property, it's a thread-local
property.  So if two requests are served by the same thread, where the
first request sets the locale to something non-default and the second
does not set it at all, the second request will use the same
non-default locale as the first one.

This is described in the i18n docs:
https://guides.rubyonrails.org/i18n.html

The homepage controller does not set the locale, and so if a user
views a Welsh content item and then the same thread is used to serve
the homepage, the homepage will use the Welsh locale.  This isn't a
new problem, but it didn't cause any issues, because Rails falls back
to the default locale (English in our case) if a translation is
missing, and we didn't have translated content for the homepage.

fe2b0b7 added those missing Welsh translations.

This commit always sets the locale to the default locale at the start
of a request.  A better option would be to use `I18n.with_locale` and
an `around_action` for those controller methods which use a different
locale, but that would be a larger refactor.