Remove uses of `has_unconfirmed_email`
This attribute is no longer relevant since we switched to using DI's SSO
service - they require users to confirm their email address as part of
the account creation journey.

These helper functions aren't being used in the account home view, so
this commit has no affect on the behaviour of appearance of this app.