Move first-time form to a separate page, and redirect to it
This commit moves the first-time form from the /sign-in/callback
endpoint to a new /sign-in/first-time endpoint, and redirects to it if
needed.

This has a couple of benefits:

1. The callback endpoint is now simpler: it either throws an error or
redirects.

2. Refreshing the form will no longer re-validate the code & state
parameters.

The downside is that we need a way of passing the session header value
from the callback endpoint to the form *without* logging the user in,
as they're not actually logged in until they submit it.  I've solved
this problem by storing the value in the header, but with a prefix,
which means account-api will be unable to decode it if the user
wanders off to another page and tries to use their account.