Flesh out signing in integration tests
This also addresses two minor points with the sessions controller:

- If the _ga parameter is present but blank, strip it from the
destination URL.  There's no point passing along an empty GA session
id.

- Gracefully handle the lack of a redirect path for a first-time user.
This is strictly speaking a bug, but it's unlikely that a user would
stumble upon this, as they would have to directly enter the URL for
www.gov.uk/account (or /sign-in/redirect), without clicking a link to
it, despite not already having an account.

These tests don't cover the special request / response header logic.
They just stub it out.  I don't think we can properly handle it
without being able to, eg, run code in between capybara receiving a
redirect response and following it (to update the request headers
based on the response).  But the headers are covered by the functional
tests for the controller.

Also, I found the need to stub out the email-alert-api calls a little
awkward, and feel that we should probably move that communication to
account-api as part of the GET /api/user behaviour, rather than doing
it in frontend.