Don't pass _ga param after first time consent page
If a user has landed on the first time consent page with a _ga parameter
then they've accepted analytics cookies on DI's domain. We always run
Google Analytics in this case, so the linker parameter will be picked
up and a new cookie will be written on this page with the client ID.
This means we don't need to pass the parameter on to the next page as
well.

If they haven't got a _ga parameter on this page, then they haven't
accepted analytics cookies on DI's domain and we have nothing to pass on
anyway.

In either case, we don't need to pass the param on to the next page in
order to get the behaviour we want.