Change popular links to use content from publishing_components
Since the header redesign popular links now appear in two places, once
on the home page and one on the header.

If these two list diverge it could be confusing for users, and updating
now has to occur in two places.

However, as the publishin components are a dependency of frontend,
content stored in the locale file here is already loaded into the app.

This means we can remove this content from the application, and load it
directly from the compoent content.

As a result there is a single place we can upload changes.