Make homepage / Welsh test less coupled to Welsh locale file
The previous iteration of this test checked that 1) English text
appeared, and 2) Welsh text did not appear.
Whilst that fulfilled the needs of the test, 2) is a bit redundant,
given that we don't render a Welsh homepage at all. It means having
to maintain a set of Welsh locales for a homepage, which never ends
up getting used.

It should be sufficient to just check that the English text displays.
I've chosen a more substanial English locale property than just
"Topics", which is the current value of `services_and_information`
and could risk being part of 'normal' page furniture (as not
everything on a Welsh page is translated).