Tighten up locale validation config
Previously we were telling Rails Translation Manager to ignore
everything under `formats.transaction`. This is too broad; we
actually only want to ignore `formats.transaction.other`.

`:other` is a keyword in Rails i18n; its usage is assumed to refer
to pluralisation rules, and thus Rails Translation Manager
complains when it is not accompanied by `:one` (in English) and
`:few`, `:zero`, `:two` and `:many` in other languages.

In such cases, we can either rename the key, or tell Rails
Translation Manager to ignore it, which we've done here.