Remove unused Start a business AB test code
This AB test was stopped in 2021 see
https://github.com/alphagov/govuk-cdn-config/pull/366

The old code was causing a test[1] to be flakey

[1] https://github.com/alphagov/frontend/blob/57cd4199d31da200757e50f653c51c2f211a3b24/test/integration/recruitment_banner_test.rb#L44

The old AB test showed an intervention banner for the B variant, so occasionally
the banner was being rendered on transaction pages. As the failing test used a
transaction fixture, the test to confirm that no banner is shown was failing.