Remove first time sign in page
When we first built the account home area, we were expecting it to be
part of `www.gov.uk` with only the authentication pages on
`account.gov.uk`. Since then, the team building the account area has
moved to the Digital Identity directorate and we've decided that all of
the central account pages should be on `account.gov.uk` and `www.gov.uk`
should be treated as any other service that's using GOV.UK Sign In as
an OIDC provider.

The first time sign in page in this app currently has part of the
account privacy policy and asks users for their consent to use cookies
and send them emails to ask for feedback. These questions need to be
moved along with the rest of the account home area to a new central
app on `account.gov.uk` and so will not be stored in GOV.UK's account
API.

This app uses the presence of answers to these questions to decide if a
user is signing in for the first time or not. If the questions are no
longer present, then this will no longer be possible. The only other
content on this page is privacy information which we are going to move
to email alert frontend.

Therefore we can remove the first time sign in page and logic in the
callback endpoint to detect if a user is signing in for the first time.