Remove account cookies and feedback page
In [[1]] we removed the first time sign up page. This was the only place
where the stored cookies and feedback data was used - to see if the
user signing in was doing so for the first time.

We no longer want to store this information in GOV.UK's account API as
we are switching GOV.UK's integration with GOV.UK Sign In to that of
any other relying party. If we do want to ask account holders for
consent about cookies and feedback then we'll create new pages in the
central `account.gov.uk` area where we can potentially pass this
information to other services they use with their account.

For now though, since we don't need the data we can remove the page
from frontend.

[1]: https://github.com/alphagov/frontend/pull/3224