Re-set the account session header on the account home page
This wasn't getting set when running the test suite, so the calls to
email alert API wasn't being stubbed properly as it didn't have the
expected session header.

Re-set the session header after getting the user information from
account API. This won't have any noticeable impact to users, but it
does mean the cookie will be applied properly in the tests.