Add optional anchor to location form results
When submitting a location form, the page scroll position would reset to the top
of the page. This wasn't an issue on pages with a small amount of content but
for pages with lots of content, it made it look as if there were no results. A
optional anchor param has been added to the place partial which can be set in
the place_controller. This will be used to generate a URL with an anchor that
will be assigned to the action of the form, so that the page will jump to the
results on submit. The id of the results container will also be set to this
anchor. As we are now conditionally setting attributes on the results section
and the form itself, these have been changed to content_tag and form_with
respectively. This was the most elegant solution to avoid having element
attributes with nil values.