Add anchor link to error form location
When submitting a location form and the submit response
was an error, the page scroll position would reset to the
top. This was an issue on pages with long content before
location forms, as it was not clear that the form had submitted
or where the error was. Now the error message can be assigned
an id, so that it can be anchor linked to.