GUA-72: Update account home redirect
Now the `GovukPersonalisation::Urls.your_account` links have all been
updated to point to DI, we can update the redirect on this side to
point to the page that will house the new account home page on the DI
side.

At the moment, this page simply redirects on to the 'manage your
account' page, but setting the GOV.UK links up like this means they'll
work properly once we do build a new account home page.