Display UR recruitment banner on pages tagged to Working browse
The number of responses hasn't been as much as we hoped, so I'd like to put the
banner out more widely. This adds the banner to all pages tagged to 'Working,
jobs and pensions' Excluding the mainstream browse page itself, as that's been
looked at in the A/B test.