Turn off font
We should serve the font files from Static so they're cached across the entirety
of GOV.UK - so the font shouldn't be included in Frontend's stylesheet.

With the font being served by each application, users need to re-download it
every time they went from a page rendered in one application to a page
rendered in another application.

Serving it from Static means that the same font URI is used in each application,
allowing the fonts to be cached and reused.