Hoist body content
The components being used outside of the `yield` in the template weren't being
added into the list of components in time - they were being added to the list
after the stylesheets were added to the `head` of the page.

Shifting the body content into another `yield` ensures that the list of
stylesheets is populated with all of the components used before application
template is run.