Fix homepage cards in IE11
Recently, a new card was added to the list of cards that appear on the
home page. The number of items in the column mixin used for IE11 layout
was not updated however which was causing a visual issue on IE11. This
has now been changed to reflect the correct number of cards and this has
fixed the visual issue.