Add LicenceTransactionPresenter
New licence presenter which represents the specialist document data
structure. We've added two methods, one to pull the body from the
details hash and secondly, we've addded the slug method which returns
the actual slug of the document. This overrides the default behaviour
[1] which just returns the base_path which we don't want for licence
transactions (vs licences) due the the base_paths changing from /:slug
-> /find-licences/:slug.

[1]: https://github.com/alphagov/frontend/blob/main/app/presenters/content_item_presenter.rb#L22-L24