Creates new licence transaction endpoint
This commit does the following:

* adds routes for the new licence base_path (/find-licences/licence')

* creates new controller and views for Specialist Publisher licence
  (Licence Transaction). This is broadly the same as the existing
  Licence controller / views but updated to respect the new `base_path`
  and data structure.

* Adds controller test

Common classes that don't need modifying are now shared between Licence
and LicenceTransaction, e.g `LicenceDetailsPresenter`.

We need the `content_item_slug` method in `LicenceTransactionController`
to ensure the correct content item is returned due to [1], as
`params[:slug]` was the content item base_path in old licence world but
now it is simply the slug, e.g “tv-licence”, whereas the content item is
located at `/find-licences/tv-licence`.  `content_item_slug` is
customisable [2] and is by a few controllers.

[1]: https://github.com/alphagov/frontend/blob/be6d2fc2ee4d3def35e460397599d135c18fc517/app/controllers/content_items_controller.rb#L25
[2]: https://github.com/alphagov/frontend/blob/main/app/controllers/content_items_controller.rb#L28-L30