Photo ID promo in CompletedTransactionController
On specific pages, we want to show a Photo ID promo. These addresses
have been added to a list which the current slug of the page will be
compared to. If the current slug is in the list, then we 'fake' a
content item promo object and replace that with the default promo
attribute of the content item of the page.

Have commented what I have done because it is a temporary change. Has
been implemented this way to be the least annoying to remove once the
campaign is over.