Make completed transaction promo partial
Refactor the promo section in completed transaction pages to use a promo
partial. The section was becoming unwieldy and even though this is part
of a temporary change it didn't seem right to continue this tech debt.

As the locale files have been refactored to use objects instead of
several seperate attributes, we can now easily call a partial using by
referencing the object named with the promo type.