Add error to electoral controller
The error for location form for the electoral controller was not
displaying when the user submitted a blank form. This was because there
was not a check for when the parameters were not null but empty. We now
check to see if a the form parameters are null (which means the user has
just loaded the page and not submitted a form) or if the form paramters
are not null (which means the user has submitted the form and so we
should them an error message).