Use the same partial for all location forms
The form being used for the electoral location was a different partial
to the one being used for all other routes. As there was only minor
differences between the two, the shared location form has been slightly
modified to accept the parameters that were different between the shared
form and the specific form for the electoral location. Combining the two
means that the same error behaviour will happen on all location forms.