Add location_error to base page partial
The intended behaviour when a form has an error is that the page title
should be updated to "Error: PAGE TITLE". This was only happening on the
Find Local Council route because it has it's own base_page partial with
the logic to change the name of the page if there is an error. The
base_page partial used by all other location forms has been updated to
have the logic for this error.