Add error summary to Simple Smart Answer
Although Simple Smart Answer would show a validation error above the
radio button in the event of an error on the form, it would not show the
error summary which should be standard on all forms. This commit adds
the error summary and also tests to check the error summary will appear
when the form has errored.