Add support for District/Unitary pairs
During merge periods, it is sometimes simpler to mark a county that will turn into a unitary authority as a unitary authority before the actual switchover period. This change allows that (previously the code assumed this could never happen, and would break if it did). If we reengineer LLM after April 2023 we can probably remove this, but we'll need to be mindful of similar future events.