Add missing attributes for find_local_council test
The content_item generated for find_local_council tests was missing two
attributes which are present in the actual content_item (title and
description) for find_local_council. This was causing the tests to fail.
These two extra attributes can now be added to the content_item used for
tests which means they are no longer filled with random data and instead
reflect the actual content_item.