Serve draft CSV previews from the draft asset host
If a user visits a CSV preview but the document is draft, we don't want
to serve this document publicly.

As a solution, we need to redirect the user to the draft assets host.
Once the user has been redirected, that URL is already behind Signon, so
we will authenticate the user.