Add missing visit to windows-1252 encoding test
This is an attempt at fixing a flakey test where the tests for the
windows-1252 encoding are passing sporadically. My understanding is that
the setup step for these tests is incomplete due to the lack of a visit
method, however these tests have passed if a previous tests' visit
method was called which left the browser in a state where the tests
would pass.