Remove homepage redirect
Previously, we were redirecting '/homepage' to '/' via:
a) publishing a special route for /homepage with a rendering application of frontend
b) redirecting these requests to '/' in the routes file in this application

This is a very old redirect, from before we had a lot of the publishing infrastructure.

We've now published an equivalent redirect in short url manager - see the list of live redirects here: https://short-url-manager.publishing.service.gov.uk/list_short_urls
This pushes the redirecting to the router level, making it consistent with other redirects on GOV.UK, and  means that we can now remove the redirect from the routes file, as well as the code to publish the content item, as this is already handled by short url manager.