update routing constraint for calendars
There are three calendar paths:

* /bank-holidays
* /when-do-the-clocks-change
* /gwyliau-banc (Welsh language version of /bank-holidays)

The routing for the first two of these was previously done based on the content items (when the content item had a schema_name of calendar, it was picked up by this constraint. We are now going to publish these content items as special routes, in order to move them into special route publisher, and remove the calendar schema (since we don't publish anything substantive in the content item). This therefore updates the routing constraint to be based on these paths, rather than the schema.