Change layout of homepage
Services and information and Featured sections are now in a grid layout.
Font size of the header of the promotion slot cards has been changed to
's', the same size as those in the Services and information section.

This is part of a live test of several incremental changes being made
to the homepage.