CSV previews to work with modern urls
Context: Currently whitehall uses a semi-hard-coded url called `legacy-url-path` to identify an asset in asset manager.
This url path looks something like `/government/uploads/system/uploads/attachment_data/*`.
While other publishing services use asset-id to identify an asset in asset manager.
We are working on an epic to make whitehall use the asset-id similar to other publishing service.

For csv-preview: Currently with Frontend preview functionality for csv, it uses legacy-url-path to fetch the asset
from asset manager but with our changes when the legacy url path is removed then frontend would have to use
the asset-id to fetch the asset.

The changes we have made will allow the controller to retrieve asset using asset-id or legacy url path based on the params its called with

Co-authored-by: Syed Ali<syed.ali@digital.cabinet-office.gov.uk>