Remove remaining multiple licensing authorities
We already removed a lot of the multiple licencing authority code in
[1]. This removes the remaining aspects. We still support the case where
multiple authorities are returned, incase licences change unexpectedly,
however as we don't expect this to happen, removing the bespoke code
simplifies the flow.

We also remove the multiple authority selection text (redirects to start
page where this was previously possible).

[1]: https://github.com/alphagov/frontend/pull/3594