Include tagline inside homepage h1
As flagged in Anika's accessibility audit, the tagline had been marked up as a paragraph tag but styled to look like a heading. Visually it was only differentiated from the "GOV.UK" text with the use of colour so it looked like a part of the h1 rather than a paragraph tag.