Make design changes to the homepage live
This does the minimum to make the homepage changes live.
Before the changes were hidden behind two feature flags, one to
restrict the changes to the integration environment, and the other to
allow reviewers to flip between both versions of the homepage.

Once we're happy that the new changes have been successfully deployed,
another PR will be opened to strip out all of the extra code required
to support two versions of the homepage.