Add a space to homepage heading h1
We recently swapped the punctuation in the new visually hidden text in the homepage heading from a colon to a dash, but didn't include a space in front of the dash. This wasn't obvious visually or when tested with screen readers since the visually hidden text inserts a space chacter before it. However if the user turns off their stylesheets, there wouldn't be a space character before the dash.